requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mixpanel==4.10.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.1
stripe==10.12.0
typing_extensions==4.12.2
urllib3==2.2.3
